+++ showcase/CHANGELOG.md
@@ -1,5 +1,43 @@
 # Changelog
 
+## [0.37.2](https://github.com/tambo-ai/tambo/compare/showcase-v0.37.1...showcase-v0.37.2) (2026-02-11)
+
+
+### Bug Fixes
+
+* **ui-registry:** support email fields in Form ([#2382](https://github.com/tambo-ai/tambo/issues/2382)) ([92a3656](https://github.com/tambo-ai/tambo/commit/92a36562e061ba7ec54d1f143193102de3a2a637))
+
+## [0.37.1](https://github.com/tambo-ai/tambo/compare/showcase-v0.37.0...showcase-v0.37.1) (2026-02-11)
+
+
+### Bug Fixes
+
+* **react-sdk:** unstrictify tool call args in the SDK ([#2361](https://github.com/tambo-ai/tambo/issues/2361)) ([6240c3c](https://github.com/tambo-ai/tambo/commit/6240c3c1b421914a3ca991c7c5ddbbd849e3bc38))
+* **showcase:** update code snippets and API tables to match current component APIs ([#2367](https://github.com/tambo-ai/tambo/issues/2367)) ([69b62a8](https://github.com/tambo-ai/tambo/commit/69b62a89491f963f30182bb45193d816aa67f037))
+
+## [0.37.0](https://github.com/tambo-ai/tambo/compare/showcase-v0.36.0...showcase-v0.37.0) (2026-02-10)
+
+
+### Features
+
+* **react-sdk:** normalize thread field from title to name ([#2332](https://github.com/tambo-ai/tambo/issues/2332)) ([b940aa4](https://github.com/tambo-ai/tambo/commit/b940aa40e769ed019b08a916fccea812d97e3ce9))
+
+
+### Bug Fixes
+
+* **ui:** message generation stage, input placeholder, canvas page in showcase ([#2358](https://github.com/tambo-ai/tambo/issues/2358)) ([3acedbb](https://github.com/tambo-ai/tambo/commit/3acedbbc32437c1b8f3d8cacdf3dc012ae1abeb6))
+
+
+### Miscellaneous Chores
+
+* bump to typescript-sdk 0.92 ([#2334](https://github.com/tambo-ai/tambo/issues/2334)) ([3767c6c](https://github.com/tambo-ai/tambo/commit/3767c6c0e12c9862afd992b6f18bc3338ea4201d))
+* **showcase:** add package description ([#2310](https://github.com/tambo-ai/tambo/issues/2310)) ([28135ee](https://github.com/tambo-ai/tambo/commit/28135ee0aac257ea36ceceba477fffd5fba54390))
+
+
+### Code Refactoring
+
+* **react-sdk:** promote V1 SDK to main export (1.0.0-rc.1) ([#2297](https://github.com/tambo-ai/tambo/issues/2297)) ([1799bce](https://github.com/tambo-ai/tambo/commit/1799bceecf412d1a4f263108dae75ddcb3fe7491))
+
 ## [0.36.0](https://github.com/tambo-ai/tambo/compare/showcase-v0.35.0...showcase-v0.36.0) (2026-02-02)